feed/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
feedparser==6.0.10
feedwerk==1.1.0
idna==3.6
loguru==0.7.2
MarkupSafe==2.1.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
urllib3==2.1.0
Werkzeug==3.0.1
